Clean up download.
