+++ crates/proto/node/src/download.rs
@@ -1,11 +1,8 @@
 use crate::platform::NodeArch;
 use crate::NodeLanguage;
-use log::debug;
-use proto_core::{
-    async_trait, color, download_from_url, Describable, Downloadable, ProtoError, Resolvable,
-};
+use proto_core::{async_trait, Downloadable, ProtoError, Resolvable};
 use std::env::consts;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 #[cfg(target_os = "macos")]
 pub fn get_archive_file_path(version: &str) -> Result<String, ProtoError> {
@@ -70,31 +67,13 @@ impl Downloadable<'_> for NodeLanguage {
             .join(get_archive_file(self.get_resolved_version())?))
     }
 
-    async fn download(&self, to_file: &Path, from_url: Option<&str>) -> Result<bool, ProtoError> {
-        if to_file.exists() {
-            debug!(target: self.get_log_target(), "Tool already downloaded, continuing");
-
-            return Ok(false);
-        }
-
+    fn get_download_url(&self) -> Result<String, ProtoError> {
         let version = self.get_resolved_version();
-        let from_url = match from_url {
-            Some(url) => url.to_owned(),
-            None => {
-                format!(
-                    "https://nodejs.org/dist/v{}/{}",
-                    version,
-                    get_archive_file(version)?
-                )
-            }
-        };
 
-        debug!(target: self.get_log_target(), "Attempting to download tool from {}", color::url(&from_url));
-
-        download_from_url(&from_url, &to_file).await?;
-
-        debug!(target: self.get_log_target(), "Successfully downloaded tool");
-
-        Ok(true)
+        Ok(format!(
+            "https://nodejs.org/dist/v{}/{}",
+            version,
+            get_archive_file(version)?
+        ))
     }
 }
